Move `app list` to `get apps`
