+++ cmd/gitops/get/cmd.go
@@ -3,6 +3,7 @@ package get
 import (
 	"github.com/go-resty/resty/v2"
 	"github.com/spf13/cobra"
+	"github.com/weaveworks/weave-gitops/cmd/gitops/get/app"
 	"github.com/weaveworks/weave-gitops/cmd/gitops/get/clusters"
 	"github.com/weaveworks/weave-gitops/cmd/gitops/get/credentials"
 	"github.com/weaveworks/weave-gitops/cmd/gitops/get/templates"
@@ -13,6 +14,9 @@ func GetCommand(endpoint *string, client *resty.Client) *cobra.Command {
 		Use:   "get",
 		Short: "Display one or many Weave GitOps resources",
 		Example: `
+# Get all applications under gitops control
+gitops get apps
+
 # Get all CAPI templates
 gitops get templates
 
@@ -23,6 +27,7 @@ gitops get credentials
 gitops get clusters`,
 	}
 
+	cmd.AddCommand(app.Cmd)
 	cmd.AddCommand(templates.TemplateCommand(endpoint, client))
 	cmd.AddCommand(credentials.CredentialCommand(endpoint, client))
 	cmd.AddCommand(clusters.ClusterCommand(endpoint, client))
